Create separate SubcellularLocationWithVizView component.
